+++ examples/example-app-router/src/components/LocaleSwitcherSelect.tsx
@@ -3,7 +3,7 @@
 import clsx from 'clsx';
 import {useParams} from 'next/navigation';
 import {ChangeEvent, ReactNode, useTransition} from 'react';
-import {useRouter, usePathname} from '../navigation';
+import {useRouter, usePathname} from '@/navigation';
 
 type Props = {
   children: ReactNode;