+++ CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 3.11.3 (2024-04-17)
+
+
+### Bug Fixes
+
+* When using domain-based routing, use `defaultLocale` of a domain instead of the top-level one in case no other locale matches better on the domain ([#1000](https://github.com/amannn/next-intl/issues/1000)) ([42988b7](https://github.com/amannn/next-intl/commit/42988b773581dad7a09af0b5cd01c958afa60865)), closes [#998](https://github.com/amannn/next-intl/issues/998)
+
+
+
+
+
+## 3.11.2 (2024-04-17)
+
+
+### Bug Fixes
+
+* Correctly detect base path at the app root `/` when using a locale prefix strategy other than `always`. This ensures the locale cookie is set correctly. ([#999](https://github.com/amannn/next-intl/issues/999)) ([1ce5988](https://github.com/amannn/next-intl/commit/1ce598893281f9789046b639b39a55fe4cca34c7)), closes [#997](https://github.com/amannn/next-intl/issues/997)
+
+
+
+
+
 ## 3.11.1 (2024-04-05)